:hammer: Moved JSON serialization in output.cpp
